backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
numpy==1.24.3
pandas==2.0.3
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
Pillow==10.0.0
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
sounddevice==0.4.6
soundfile==0.12.1
python-dotenv==1.0.0
langchain==0.0.350
chromadb==0.4.18 